requirements.txt
# typing
ccxt
pytz
pandas
datetime
loguru
facade
pyyaml
schedule
pyarrow